+++ src/components/board/__tests__/ColumnResize.test.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 import { render, fireEvent } from "@testing-library/react";
-import { BoardProvider } from "../../board/BoardProvider";
-import { ClipboardProvider } from "../../board/ClipboardProvider";
-import { Column } from "./Column";
-import { ThemeContext } from "../../theme/ThemeContext";
-import type { ThemeContextValue } from "../../theme/types";
+import { BoardProvider } from "../../../board/BoardProvider";
+import { ClipboardProvider } from "../../../board/ClipboardProvider";
+import { Column } from "../Column";
+import { ThemeContext } from "../../../theme/ThemeContext";
+import type { ThemeContextValue } from "../../../theme/types";
 
 function ThemeProviderWithColumnResize({
   children,